+++ src/types/index.ts
@@ -3,3 +3,7 @@ export type post = {
 	status_message: string;
 	success: boolean;
 };
+
+export type DELETE = {
+	success: boolean;
+};